+++ lua/configs/ui/themes/gruverboxer-material.lua
@@ -1,4 +1,0 @@
-return {
-    "BrunoCiccarino/gruverboxer-material",
-    lazy = true
-}